+++ display_transaction.php
@@ -41,10 +41,10 @@ if($transactions_results){
             
             // determine color of number and current total balance
             if($transaction_type == "Deposit"){
-                $total_balance += (int)$transaction_amount;
+                $total_balance += (float)$transaction_amount;
                 $transaction_amount = "<font color='blue'> $transaction_amount </font>";
             } else {
-                $total_balance -= (int)$transaction_amount;
+                $total_balance -= (float)$transaction_amount;
                 $transaction_amount = "<font color='red'> -$transaction_amount </font>";
             }
