+++ search_transaction.php
@@ -46,10 +46,10 @@ if($search_results){
             
             // determine color of number and current total balance
             if($transaction_type == "Deposit"){
-                $total_balance += (int)$transaction_amount;
+                $total_balance += (float)$transaction_amount;
                 $transaction_amount = "<font color='blue'> $transaction_amount </font>";
             } else {
-                $total_balance -= (int)$transaction_amount;
+                $total_balance -= (float)$transaction_amount;
                 $transaction_amount = "<font color='red'> -$transaction_amount </font>";
             }
